requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.28
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
boto3==1.34.0
botocore==1.34.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
PyJWT==2.10.1
loguru==0.7.3
chardet==5.2.0